rag/requirements.txt
# Core dependencies
langchain>=0.0.267
llama-index>=0.8.4
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
torch>=2.0.0
transformers>=4.30.2

# Document processing
pypdf>=3.15.1
python-docx>=0.8.11
beautifulsoup4>=4.12.2
html2text>=2020.1.16

# Utilities
numpy>=1.24.3
pandas>=2.0.3
tqdm>=4.65.0
pydantic>=2.0.3
python-dotenv>=1.0.0

# Vietnamese language support
underthesea>=6.0.0  # Vietnamese NLP toolkit

# Evaluation and monitoring
scikit-learn>=1.3.0
matplotlib>=3.7.2